Canonicalize tests to @safetestset for isolation
Convert the plain-`@testset` independent test units to `@safetestset`
so each runs in its own freshly-generated module (world-age safety and
isolation between units), matching the canonical OrdinaryDiffEq structure.

- runtests.jl: the three independent units (`_is_valid_macro_expr`,
  `_get_symbols`, `Tests for TestModule1`) and the allocation tests are
  now `@safetestset "X" begin include("x.jl") end`, kept inside the outer
  `@testset "SciMLPublic.jl package"` grouping. `TestModule1` and its
  `_public_names` helper move into the extracted file so the `@public`
  package macro resolves after `using SciMLPublic` runs (include()
  evaluates statements sequentially).
- New self-contained files: macro_expr_tests.jl, get_symbols_tests.jl,
  testmodule1_tests.jl (each carries its own `using SciMLPublic; using Test`).
- qa/qa.jl: `Aqua`/`JET` units converted to `@safetestset`, each with its
  own imports; `@test_broken` preserved verbatim.
- Add SafeTestsets to the main Project.toml ([extras]/[targets].test/[compat])
  and to test/qa/Project.toml ([deps]/[compat]).

Behavior-preserving: same GROUP dispatch, same tests, same assertions.
Verified locally on Julia 1.11: GROUP=Core 25/25 pass; GROUP=QA Aqua 7
pass + 1 broken (preserved), JET 1 pass.
