+++ Project.toml
@@ -4,6 +4,7 @@ authors = ["The Julia Project", "contributors"]
 version = "1.0.1"
 
 [compat]
+SafeTestsets = "0.1, 1"
 Test = "1.10"
 julia = "1.10"
 
@@ -11,6 +12,7 @@ julia = "1.10"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 
 [targets]
-test = ["Test", "Aqua", "JET"]
+test = ["Test", "Aqua", "JET", "SafeTestsets"]
